+++ components/ui/LoadingIndicator.vue
@@ -1,7 +1,7 @@
 <template>
   <div class="w-40">
     <div class="bg-bg border border-border py-2 px-5 rounded-lg flex items-center flex-col box-shadow-md">
-      <div class="loader-dots block relative w-20 h-5 mt-2">
+      <div class="loader-dots tv-focus-dots block relative w-20 h-5 mt-2">
         <div class="absolute top-0 mt-1 w-3 h-3 rounded-full bg-green-500"></div>
         <div class="absolute top-0 mt-1 w-3 h-3 rounded-full bg-green-500"></div>
         <div class="absolute top-0 mt-1 w-3 h-3 rounded-full bg-green-500"></div>